[AMI] Refactor actions
