+++ asterisk-java-all/src/test/java/org/asteriskjava/manager/internal/ManagerWriterMock.java
@@ -17,15 +17,11 @@
 package org.asteriskjava.manager.internal;
 
 import org.asteriskjava.AsteriskVersion;
-import org.asteriskjava.ami.action.AuthType;
-import org.asteriskjava.ami.action.ChallengeAction;
-import org.asteriskjava.ami.action.LoginAction;
-import org.asteriskjava.ami.action.ManagerAction;
+import org.asteriskjava.ami.action.*;
 import org.asteriskjava.ami.action.response.ChallengeResponse;
 import org.asteriskjava.ami.action.response.ManagerResponse;
 import org.asteriskjava.ami.action.response.ResponseType;
 import org.asteriskjava.core.socket.SocketConnectionAdapter;
-import org.asteriskjava.manager.action.LogoffAction;
 import org.asteriskjava.manager.event.ProtocolIdentifierReceivedEvent;
 import org.asteriskjava.manager.response.ManagerError;
 import org.asteriskjava.util.DateUtil;
