+++ asterisk-java-ami/src/main/java/org/asteriskjava/ami/action/response/ResponseType.java
@@ -24,4 +24,5 @@ package org.asteriskjava.ami.action.response;
 public enum ResponseType {
     Success,
     Error,
+    Goodbye,
 }
